# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Dare To Accept API",
    "version": "1.0.0",
    "description": "API documentation for Dare To Accept platform."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/campaigns/{campaignId}/payout": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Payout campaign funds",
        "description": "Allows the campaign owner or admin to request a payout of campaign funds.",
        "operationId": "payoutCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "type": "string",
            "description": "The ID of the campaign to payout."
          }
        ],
        "responses": {
          "200": {
            "description": "Payout successful"
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found"
          }
        }
      }
    },
...
    "/challenges": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Create a new challenge",
        "description": "Allows a user to create a new challenge.",
        "operationId": "createChallenge",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChallengeRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Challenge created successfully",
            "schema": {
              "$ref": "#/definitions/Challenge"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Error creating challenge"
          }
        }
      }
    },
    "/challenges/{challengeId}/join": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Join a challenge",
        "description": "Allows a user to join a specific challenge.",
        "operationId": "joinChallenge",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge to join."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JoinChallengeRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Joined challenge successfully",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          },
          "400": {
            "description": "Already joined or invalid request"
          },
          "404": {
            "description": "Challenge not found"
          }
        }
      }
    },
    "/challenges/{challengeId}/participants/{participantId}/out": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Admit defeat in a challenge",
        "description": "Allows a participant to admit defeat in a specific challenge.",
        "operationId": "admitDefeat",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "string",
            "description": "The ID of the participant admitting defeat."
          }
        ],
        "responses": {
          "200": {
            "description": "Defeat admitted successfully",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          },
          "400": {
            "description": "Cannot admit defeat in current status"
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
    "/challenges/{challengeId}/participants/{participantId}/dispute": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Dispute result of a challenge",
        "description": "Allows a participant to dispute the result of a specific challenge.",
        "operationId": "disputeResult",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "string",
            "description": "The ID of the participant disputing the result."
          }
        ],
        "responses": {
          "200": {
            "description": "Dispute triggered successfully",
            "schema": {
              "$ref": "#/definitions/Challenge"
            }
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
    "/challenges/{challengeId}/confirm-winner": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Confirm winner of a challenge",
        "description": "Allows the creator of a challenge to confirm the winner.",
        "operationId": "confirmWinner",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConfirmWinnerRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Winner confirmed successfully",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          },
          "403": {
            "description": "Only the creator can confirm the winner"
          },
          "404": {
            "description": "Challenge or winner not found"
          }
        }
      }
    },
    "/challenges/{challengeId}/vote": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Vote for a winner during dispute phase",
        "description": "Allows a user to vote for a winner during the dispute phase of a challenge.",
        "operationId": "voteForWinner",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VoteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vote recorded successfully",
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          },
          "400": {
            "description": "Already voted or voting not open"
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
...
  },
...
}